Answer "read this from JSON" in one place the compiler owns
`souther run` composed its own decoders for the shapes no class is
emitted for — a list, a set, a map, a primitive, a temporal — and its own
encoders to write the answer back. `DecoderPathAgreementTest` records what
that costs: three separate implementations answer "read a value of type T
from the outside", they are written independently, and issue #97 was one
of them drifting from the others.

Move that composition to `souther.compiler.JsonBoundary`, beside the
generated codecs it has to agree with. `Runner` keeps what only a command
line knows — where a value sat in `--input`, what to call the failure, how
to print it — and the boundary answers with a `Read` and with the
reflective failure it got, saying nothing to a reader.

`Runner` reaches nine fewer of the compiler's types for it, and no longer
reaches `codegen` or `types` at all. What is left is the signature it
picks a behavior by.

`DecoderPathAgreementTest` drives `JsonBoundary` rather than `souther run`,
so the table that pins the three readers together no longer needs the
command line to reach the third. Its three key tests read the refusal
itself — the issues and the paths they stand at — instead of the message
`souther run` would have made of them.
